qulity to i64

+++ src/entities.rs
@@ -381,8 +381,8 @@ pub struct VideoUrl {
     pub result: String,
     #[serde(default = "default_string")]
     pub message: String,
-    #[serde(default = "default_i32")]
-    pub quality: i32,
+    #[serde(default = "default_i64")]
+    pub quality: i64,
     #[serde(default = "default_string")]
     pub format: String,
     #[serde(default = "default_i32")]
